requirements.txt
opencv-python~=4.10.0.84
numpy~=2.2.0
scikit-learn~=1.5.2
matplotlib
pandas
seaborn
python-multipart
Jinja2
shutils
fastapi==0.95.1
uvicorn==0.22.0
mysql-connector-python==8.0.29